+++ bee-api/bee-rest-api/CHANGELOG.md
@@ -47,6 +47,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `AliasIdDto`;
 - `TokenSchemeDto`;
 - `NftIdDto`;
+- `RegularTransactionEssenceDto::network_id`;
 
 ### Changed